16.Image Producer/ArtifexGen/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.0
diffusers==0.21.4
filelock==3.12.4
Flask==3.0.0
fonttools==4.43.0
fsspec==2023.9.2
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
packaging==23.2
Pillow==10.0.1
psutil==5.9.5
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.3.3
six==1.16.0
sympy==1.12
tokenizers==0.14.0
torch==2.1.0
tqdm==4.66.1
transformers==4.34.0
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
zipp==3.17.0
accelerate==0.24.1
